requirements.txt
asgiref==3.5.2
Django==4.1.5
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7

djangorestframework~=3.14.0